Functions

CryptoUtils

Number to bytes conversion

CryptoUtils.b2n — Function.

b2n(str::String) -> BigInt

Converts a byte-string to a number, converting the string from base 256 to base 10.

julia> b2n("Hello world!")
22405534230753963835153736737

source

CryptoUtils.n2b — Function.

n2b(n::Integer) -> String

Converts a number to its bytes representation, effectively writing the number in base 256, and returning the corresponding bytes.

julia> n2b(22405534230753963835153736737)
"Hello world!"

source

Prime numbers

CryptoUtils.random_prime — Function.

random_prime(bitsize::Integer) -> BigInt

Return a random prime with bitsize bits.

julia> random_prime(42)
2458636110727

source

CryptoUtils.safe_prime — Function.

safe_prime(bitsize::Integer) -> BigInt

Return a random safe-prime q of the form q = 2 * p + 1 where p is also a prime number. The returning prime number has bitsize bits.

julia> safe_prime(10)
1439

source

CryptoUtils.tower_two_prime — Function.

tower_two_prime(bitsize::Integer, tower_len::Integer) -> BigInt

Return a random prime of the form 2^towerlen * q + 1 with bitsize bits and where q is also a prime.

julia> tower_two_prime(22, 6)
2362433

source

CryptoUtils.get_first_primes — Function.

get_first_primes(k::Integer) -> Collection

Output the first k prime numbers

julia> get_first_primes(10)
10-element Array{Int64,1}:
  2
  3
  5
  7
 11
 13
 17
 19
 23
 29

source

Number theory

CryptoUtils.find_quadratic_non_residue — Function.

find_quadratic_non_residue(p::Integer)

Return a random number R which has no square root mod p, i.e., x^2 == R mod p has no solutions.

source

CryptoUtils.hoc_sqrt — Function.

hoc_sqrt(a::Integer, p::Integer)

Algorithm from Handbook of cryptography, Koblitz pp 48-49. Finds a solution to x^2 == a mod p.

It assumes such solution exists.

Running time highly depends on |alpha|, assuming p-1 = 2^alpha * s, for an odd s.

source

CryptoUtils.is_quadratic_residue — Function.

is_quadratic_residue(a::Integer, p::Integer) -> Bool

Return true or false depending on wheter a is a quadratic residue mod p.

That is, it checks if x^2 == a mod p has solutions.

source

CryptoUtils.jacobi — Function.

jacobi(n::Integer, k::Integer)

Return the Jacobi symbol of n, k.

k should be an odd number.

source

CryptoUtils.legendre — Function.

legendre(a::Integer, p::Integer)

Return the Legendre symbol of (a, p).

p should be an odd prime number.

source

CryptoUtils.sqrt_mod_prime — Function.

sqrt_mod_prime(a::Integer, p::Integer) -> Integer

Solves x^2 == a mod p and returns one of the square roots r. The other root is p - r. If there are no solutions, throws an exception.

julia> sqrt_mod_prime(33^2, 73)
33

source

CryptoUtils.tonelli_shanks — Function.

tonelli_shanks(a::Integer, p::Integer)

Implements the Tonelli Shanks algorithm for computing square roots modulo a prime number.

It assumes such square roots exist.

source

Cryptography

CryptoUtils.factor_with_ed — Function.

factor_with_ed(n::Integer, e::Integer, d::Integer) -> (Integer, Integer)

Factors n = p*q given (e, d) such that e*d = 1 mod phi(n) Stinson page 204 - algorithm 5.10

source